requirements.txt
Flask-Assets==0.10
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gunicorn==19.0.0
httplib2==0.9
itsdangerous==0.24
python-instagram==1.3.0
simplejson==3.6.5
six==1.8.0
webassets==0.10.1
